refactor(gsd): remove most runtime auto-worktree barrel imports
