+++ src/resources/extensions/gsd/auto-worktree-resource-version.ts
@@ -1,0 +1,9 @@
+// gsd-pi — Narrow auto-worktree resource version seam.
+//
+// Keeps resource freshness callers off the legacy auto-worktree compatibility
+// barrel while extraction continues.
+
+export {
+  checkResourcesStale,
+  readResourceVersion,
+} from "./auto-worktree.js";
