+++ src/resources/extensions/gsd/auto-worktree-runtime-cleanup.ts
@@ -3,4 +3,7 @@
 // Keeps runtime cleanup callers off the legacy auto-worktree compatibility
 // barrel while extraction continues.
 
-export { cleanStaleRuntimeUnits } from "./auto-worktree.js";
+export {
+  cleanStaleRuntimeUnits,
+  escapeStaleWorktree,
+} from "./auto-worktree.js";
